% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFMfunction.R
\name{RFMfunction}
\alias{RFMfunction}
\title{RFMfunction
Calculate a weighted RMF score, this is a Description}
\usage{
RFMfunction(data, weight_recency = 1, weight_frequency = 1,
  weight_monetary = 1)
}
\arguments{
\item{data}{- a data table}

\item{weight_recency}{- weight of recency}

\item{weiht_frequency}{- weight of frequency}
}
\value{
Returns a data.table containing ... xy ...
score this is the second line. thank you for smoking.
}
\description{
RFMfunction
Calculate a weighted RMF score, this is a Description
}
\details{
mmmmmmmmmmm
}
